requirements.txt
asgiref==3.8.1
Django==5.2.3
pysqlite3-binary==0.5.4
sqlparse==0.5.3
